Add files via upload
The synapse identification pipeline in CellProfiler is referenced in the code. It highlights ring areas around each cell and then looks for colocalization of each synaptic marker within that ring. I've added a file titled "SynapseImage" which can be uploaded to the pipeline (CellProfiler must be downloaded first) and tested.